crud speaker

+++ src/components/speaker/speaker.vue
@@ -2,24 +2,24 @@
   <v-container>
 
     <div class="mt-5 mb-5">
-      <h1>Daftar Media Partner</h1>
+      <h1>Daftar Speaker</h1>
     </div>
 
     <div>
-      <button class="btn save-btn" v-b-modal.modal-1>&#43; Tambah Media Partner</button>
+      <button class="btn save-btn mb-5" v-b-modal.modal-1>&#43; Tambah Speaker</button>
     </div>
 
     <table class="table table-hover">
         <thead class="table-borderless">
             <tr class="tr-top">
                 <th scope="col">No.</th>
-                <th scope="col">Nama Media Partner</th>
+                <th scope="col">Nama Speaker</th>
             </tr>
         </thead>
         <tbody class="tbody">
-            <tr v-for="(media,index) in speakerList" :key="media.id"  v-b-modal.modal-2 @click="click(media.id)"  class="content">
+            <tr v-for="(speaker,index) in speakerList" :key="speaker.id"  v-b-modal.modal-2 @click="click(speaker.id)"  class="content">
                 <th scope=row class="th-bottom">{{index+1}}</th>
-                <td>{{media.name}}</td>
+                <td>{{speaker.name}}</td>
                 
             </tr>
         </tbody>
@@ -27,19 +27,29 @@
 
     
 
-    <b-modal size="lg" ref="modalAdd" id="modal-1" title="Tambah Media Partner" v-bind:hide-footer="true">
+    <b-modal size="lg" ref="modalAdd" id="modal-1" title="Tambah Speaker" v-bind:hide-footer="true">
       <div class="card">
-        <div class="card-header"><strong class="labelForm">Formulir Tambah Media Partner</strong></div>
+        <div class="card-header"><strong class="labelForm">Formulir Tambah Speaker</strong></div>
         <div class="card-body">
           <b-form @submit.prevent="validateAndSubmit">
             <!-- Field Nama Event -->
             <b-form-group>
-              <div class="mb-2 label"><strong class="labelForm">Nama Media Partner</strong></div>
+              <div class="mb-2 label"><strong class="labelForm">Nama Speaker</strong></div>
               <b-form-input
                 class="form-control"
-                v-model="media.name"
+                v-model="speaker.name"
                 id="eventName"
-                placeholder="Nama Media Partner"></b-form-input>
+                placeholder="Nama Speaker"></b-form-input>
+            </b-form-group>
+
+            <!-- Field Nama Event -->
+            <b-form-group>
+              <div class="mb-2 label"><strong class="labelForm">Profesi</strong></div>
+              <b-form-input
+                class="form-control"
+                v-model="speaker.profesi"
+                id="eventName"
+                placeholder="Profesi Speaker"></b-form-input>
             </b-form-group>
 
             <!-- Field tanggal event -->
@@ -86,7 +96,7 @@
             </b-form-group> -->
                
             <div class="btn-group">
-              <button type="submit" class="btn save-btn mr-2" :disabled=isDisable()>Tambah Media Partner</button>
+              <button type="submit" class="btn save-btn mr-2" :disabled=isDisable()>Tambah Speaker</button>
               <button class="btn btn-light" @click="batal">Batal</button>
             </div>
           </b-form>
@@ -99,7 +109,7 @@
     <b-modal size="lg" ref="modalOk" hide-footer>
         <div class="container">
             <div class="d-block text-center">
-              <h4>Media Partner berhasil ditambahkan</h4>
+              <h4>Speaker berhasil ditambahkan</h4>
             </div>
         </div>
     </b-modal>
@@ -107,24 +117,33 @@
     <!-- Modal Edit Event -->
     <b-modal size="lg" ref="modalEdit" id="modal-2" title="Edit Media Partner" v-bind:hide-footer="true">
       <div class="card">
-        <div class="card-header"><strong class="labelForm">Formulir Ubah Media Partner</strong></div>
+        <div class="card-header"><strong class="labelForm">Formulir Ubah Speaker</strong></div>
         <div class="card-body">
           <b-form @submit.prevent="validateAndSubmitEdit">
             <b-img-lazy class="mb-4" :src="headerPhotos"></b-img-lazy>
             <!-- Field Nama Event -->
             <b-form-group>
-              <div class="mb-2 label"><strong class="labelForm">Nama Media Partner</strong></div>
+              <div class="mb-2 label"><strong class="labelForm">Nama Speaker</strong></div>
               <b-form-input
                 class="form-control"
-                v-model="targetMedia.name"
+                v-model="targetSpeaker.name"
                 id="eventName"
-                placeholder="Nama Event"></b-form-input>
+                placeholder="Nama Speaker"></b-form-input>
+            </b-form-group>
+
+            <b-form-group>
+              <div class="mb-2 label"><strong class="labelForm">Profesi Speaker</strong></div>
+              <b-form-input
+                class="form-control"
+                v-model="targetSpeaker.profesi"
+                id="eventName"
+                placeholder="Profesi Speaker"></b-form-input>
             </b-form-group>
 
             <!-- Field tanggal event -->
             <!-- <b-form-group>
                 <div class="mb-2 label"><strong class="labelForm">Tanggal Event</strong></div>
-                <b-form-datepicker id="datepickerEnd-invalid" :state="birthdateValidEdit" class="mb-2" v-model="targetMedia.date"></b-form-datepicker>
+                <b-form-datepicker id="datepickerEnd-invalid" :state="birthdateValidEdit" class="mb-2" v-model="targetSpeaker.date"></b-form-datepicker>
                 
                 <b-form-invalid-feedback id="input-live-feedback-end">
                   Tanggal yang dimasukan tidak valid
@@ -155,7 +174,7 @@
               <div class="mb-2 label"><strong class="labelForm">Deskripsi Event</strong></div>
               <b-form-textarea
                 id="textarea"
-                v-model="targetMedia.description"
+                v-model="targetSpeaker.description"
                 placeholder="Masukan deskripsi event"
                 rows="3"
                 max-rows="6"
@@ -163,10 +182,10 @@
             </b-form-group> -->
                
 
-            <button type="submit" style="color:white" class="btn btn-block save-btn mb-2" :disabled=isDisableEdit()>Ubah Media Partner</button>
+            <button type="submit" style="color:white" class="btn btn-block save-btn mb-2" :disabled=isDisableEdit()>Ubah Speaker</button>
             
           </b-form>
-          <button type="submit" class="btn btn-block btn-light mr-2" v-b-modal.modal-del >Hapus Media Partner</button>
+          <button type="submit" class="btn btn-block btn-light mr-2" v-b-modal.modal-del >Hapus Speaker</button>
 
         </div>
 
@@ -176,7 +195,7 @@
     <b-modal size="lg" ref="modalOk" hide-footer>
         <div class="container">
             <div class="d-block text-center">
-              <h4>Media partner berhasil ditambahkan</h4>
+              <h4>Speaker berhasil ditambahkan</h4>
             </div>
         </div>
     </b-modal>
@@ -185,14 +204,14 @@
     <b-modal size="lg" ref="modalOkEdit" hide-footer>
         <div class="container">
             <div class="d-block text-center">
-              <h4>Media Partner berhasil diubah</h4>
+              <h4>Speaker berhasil diubah</h4>
             </div>
         </div>
     </b-modal>
 
     <b-modal size="lg" ref="modalDelete" id="modal-del" hide-footer>
        <div class="detail">
-            <p class="title">Media partner ini akan dihapus ? </p>
+            <p class="title">Speaker ini akan dihapus ? </p>
             <hr>
             <div class="btn-group">
                 <button type="submit" class="btn btn-danger mr-2" @click="deleteEvent">Hapus</button>
@@ -215,14 +234,16 @@ import {db, storage} from "../../firebase/firebase"
             speaker:{
               name:"",
               profesi:"",
+              photo :"",
             },
             speakerList : [],
             headerPhotos:"",
-            mediaId : "",
-            targetMedia: {},
+            speakerId : "",
+            targetSpeaker: {},
             imageData : null,
             picture : null,
             uploadValue :0,
+            avatarImg : ""
         };
     },
 
@@ -235,13 +256,13 @@ import {db, storage} from "../../firebase/firebase"
         this.imageData = event.target.files[0];
       },
       loadMedia() {
-        var mediaRef = db.collection('media-partner');
-        mediaRef.onSnapshot(snap => {
+        var speakerRef = db.collection('speakers');
+        speakerRef.onSnapshot(snap => {
           this.speakerList = [];
           snap.forEach(doc => {
-            var media =doc.data();
-            media.id = doc.id;
-            this.speakerList.push(media);
+            var speaker =doc.data();
+            speaker.id = doc.id;
+            this.speakerList.push(speaker);
             
           })
          
@@ -249,16 +270,24 @@ import {db, storage} from "../../firebase/firebase"
         // var storageRef = storage.ref().child('images').child('546FO3IJN82I2LWLV87YV.mp4_snapshot_21.14_[2019.06.28_22.26.16].png');
         // console.log(storageRef.location.path)
       },
-      click(mediaId){
-        this.mediaId = mediaId;
+      click(speakerId){
+        this.speakerId = speakerId;
 
-        db.collection('media-partner').doc(this.mediaId).get().then(doc => {
+        db.collection('speakers').doc(this.speakerId).get().then(doc => {
           // console.log(doc.data());
-          this.targetMedia = doc.data();
-          var gsReference = storage.refFromURL(this.targetMedia.photo);  
-          gsReference.getDownloadURL().then(link => {
-              this.headerPhotos = link;
-          })
+          this.targetSpeaker = doc.data();
+          if(this.targetSpeaker.photo != null){
+            var gsReference = storage.refFromURL(this.targetSpeaker.photo);  
+            gsReference.getDownloadURL().then(link => {
+                this.headerPhotos = link;
+            })
+          }
+          else{
+              var gsReferences = storage.refFromURL('gs://kmmb-website.appspot.com/images/speakers/avatar.png');  
+            gsReferences.getDownloadURL().then(link => {
+                this.headerPhotos = link;
+            })
+          }
           
         });
         
@@ -276,26 +305,50 @@ import {db, storage} from "../../firebase/firebase"
             e.preventDefault();
             this.errors = [];
             
-            if(this.errors.length === 0){
-              this.picture=null;
-              const storageRef = storage.ref().child(`images/media-partner/${this.imageData.name}`).put(this.imageData);
-              storageRef.on(`state_changed`, snapshot => {
-                this.uploadValue = (snapshot.bytesTransferred/snapshot.totalBytes)*100;
-              }, error =>{console.log(error.message)}, () => {
-                this.uploadValue=100;
-                storageRef.snapshot.ref.getDownloadURL().then((url) => {
-                  this.picture = url
-                  db.collection('media-partner').add({
-                      name:this.media.name,
-                      photo : `gs://kmmb-website.appspot.com/images/media-partner/${this.imageData.name}`,
-                  }).then(() => {
-                      this.openModal();
-                      this.uploadValue=0;
-                      
-                    });
-                })
-              }) 
+             if(this.imageData != null){
+               
+                if(this.errors.length === 0){
               
+                  const storageRef = storage.ref().child(`images/speakers/${this.imageData.name}`).put(this.imageData);
+                  storageRef.on(`state_changed`, snapshot => {
+                    this.uploadValue = (snapshot.bytesTransferred/snapshot.totalBytes)*100;
+                  }, error =>{console.log(error.message)}, () => {
+                    this.uploadValue=100;
+                    storageRef.snapshot.ref.getDownloadURL().then((url) => {
+                      this.picture = url
+                      db.collection('speakers').add({
+                          name : this.speaker.name,
+                          profesi : this.speaker.profesi,
+                          photo : `gs://kmmb-website.appspot.com/images/speakers/${this.imageData.name}`
+                      }).then(() => {
+                        this.openModal()
+                        this.uploadValue=0;
+                        this.speaker = {}
+                      })
+                    .then(() => {
+                        this.openModal()
+                        this.uploadValue=0;
+                        this.speaker = {}
+                      });
+                    })
+                  }) 
+                  
+                }
+            }
+            else{
+              db.collection('speakers').add({
+                    name : this.speaker.name,
+                    profesi : this.speaker.profesi,
+                }).then(() => {
+                  this.openModal()
+                  this.uploadValue=0;
+                  this.speaker = {}
+                })
+              .then(() => {
+                  this.openModal()
+                  this.uploadValue=0;
+                  this.speaker = {}
+                });
             }
             
 
@@ -304,22 +357,23 @@ import {db, storage} from "../../firebase/firebase"
             e.preventDefault();
             this.errors = [];
             if(this.imageData != null){
-                var filename = this.targetMedia.photo.split('/').pop().split('#')[0].split('?')[0];
-                var target = storage.ref().child('images/media-partner/'+ filename)
+                var filename = this.targetSpeaker.photo.split('/').pop().split('#')[0].split('?')[0];
+                var target = storage.ref().child('images/speakers/'+ filename)
                 target.delete();
 
                 if(this.errors.length === 0){
               
-                  const storageRef = storage.ref().child(`images/media-partner/${this.imageData.name}`).put(this.imageData);
+                  const storageRef = storage.ref().child(`images/speakers/${this.imageData.name}`).put(this.imageData);
                   storageRef.on(`state_changed`, snapshot => {
                     this.uploadValue = (snapshot.bytesTransferred/snapshot.totalBytes)*100;
                   }, error =>{console.log(error.message)}, () => {
                     this.uploadValue=100;
                     storageRef.snapshot.ref.getDownloadURL().then((url) => {
                       this.picture = url
-                      db.collection('media-partner').doc(this.mediaId).update({
-                          name : this.targetMedia.name,
-                          photo : `gs://kmmb-website.appspot.com/images/media-partner/${this.imageData.name}`
+                      db.collection('speakers').doc(this.speakerId).update({
+                          name : this.targetSpeaker.name,
+                          profesi : this.targetSpeaker.profesi,
+                          photo : `gs://kmmb-website.appspot.com/images/speakers/${this.imageData.name}`
                       }).then(() => {
                         this.openModalSusksesEdit()
                         this.uploadValue=0;
@@ -334,8 +388,9 @@ import {db, storage} from "../../firebase/firebase"
                 }
             }
             else{
-              db.collection('media-partner').doc(this.mediaId).update({
-                    name : this.targetMedia.name,
+              db.collection('speakers').doc(this.speakerId).update({
+                    name : this.targetSpeaker.name,
+                    profesi : this.targetSpeaker.profesi,
 
                 }).then(() => {
                   this.openModalSusksesEdit()
@@ -349,11 +404,15 @@ import {db, storage} from "../../firebase/firebase"
            
         },
         deleteEvent(){
-          var filename = this.targetMedia.photo.split('/').pop().split('#')[0].split('?')[0];
-          var target = storage.ref().child('images/media-partner/'+ filename)
+            if(this.targetSpeaker.photo != null){
+                var filename = this.targetSpeaker.photo.split('/').pop().split('#')[0].split('?')[0];
+                var target = storage.ref().child('images/speakers/'+ filename)
+                target.delete();
 
-          target.delete();
-            db.collection('media-partner').doc(this.mediaId).delete().then(() =>{
+            }
+          
+          
+            db.collection('speakers').doc(this.speakerId).delete().then(() =>{
                 this.hideModal();
             })
         },
@@ -381,28 +440,30 @@ import {db, storage} from "../../firebase/firebase"
           this.$refs['modalOk'].hide();
         },
         isDisable(){
-            if(this.media.name == ""){return true;}
+            if(this.speaker.name == ""){return true;}
+            if(this.speaker.profesi == ""){return true;}
  
             return false;
         },
 
         isDisableEdit(){
-            if(this.targetMedia.name == ""){return true;}
+            if(this.targetSpeaker.name == ""){return true;}
+            if(this.targetSpeaker.profesi == ""){return true;}
  
             return false;
         },
-        openModalEdit(mediaId){
+        openModalEdit(speakerId){
           
           this.$refs['modalEdit'].show();
-          this.mediaId = mediaId;
-          db.collection('media-partner').doc(mediaId).get().then(doc => {
-            this.targetMedia = doc.data();
+          this.speakerId = speakerId;
+          db.collection('speakers').doc(speakerId).get().then(doc => {
+            this.targetSpeaker = doc.data();
           })
         },
-        openModalDelete(mediaId){
+        openModalDelete(speakerId){
           
           this.$refs['modalDelete'].show();
-          this.mediaId = mediaId;
+          this.speakerId = speakerId;
         },
     },
     created(){
